Moved UserProvider to page wrap level, added token clearing logic
