+++ src/contexts/user-context.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { getUserProfileFromJWT } from '../clients/auth.js';
+import { getUserProfileFromJWT } from '../services/auth.js';
+import { useLocation } from '@reach/router';
 // import { isBrowser } from '../utils/is-browser';
 
 const defaultContextValue = {
@@ -11,13 +12,16 @@ const UserContext = React.createContext(defaultContextValue);
 
 const UserProvider = ({ children = {} }) => {
   const [userData, setUserData] = useState();
+
+  const location = useLocation();
+
   useEffect(() => {
     (async () => {
       const userProfile = await getUserProfileFromJWT();
       setUserData(userProfile);
       console.log(userProfile);
     })();
-  }, []);
+  }, [location]);
   return <UserContext.Provider value={{ userData, setUserData }}>{children}</UserContext.Provider>;
 };
